requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
passlib[bcrypt]
python-jose
httpx
python-dotenv